requirements.txt
agate==1.9.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0

apispec==6.3.0
apprise==1.9.2
argcomplete==3.2.1
asgi-lifespan==2.1.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
azure-core==1.32.0
azure-storage-blob==12.25.1
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
bleach==6.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
coolname==2.2.0
cron_descriptor==1.4.0
croniter==2.0.1
cryptography==44.0.2
daff==1.3.46
dateparser==1.2.1
dbt-adapters==1.14.3
dbt-common==1.16.0
dbt-core==1.9.3
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
dbt-snowflake==1.9.2
debugpy==1.8.13
decorator==5.2.1
deepdiff==7.0.1
Deprecated==1.2.18
dill==0.3.1.1
distlib==0.3.9
dnspython==2.4.2
docker==7.1.0
docutils==0.20.1
exceptiongroup==1.2.2
executing==2.2.0
Faker==37.1.0
fastapi==0.115.12
filelock==3.18.0
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2025.3.0
google-re2==1.1
googleapis-common-protos==1.62.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.1
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.2
hyperframe==6.1.0
idna==3.10
importlib-resources==6.1.1
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.7.4.2
kagglehub==0.3.10
keyring==25.6.0
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==3.14
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
ordered-set==4.1.0
orjson==3.10.16
packaging==23.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.3.0
prefect==3.2.15
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.15.0
python-slugify==8.0.4
python-socks==2.7.1
pytimeparse==1.1.8
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.3.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.9.4
rich-argparse==1.4.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
SecretStorage==3.3.3
setproctitle==1.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.14.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.2
ujson==5.10.0
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.3.0
uv==0.6.10
uvicorn==0.34.0
virtualenv==20.29.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.17.2
WTForms==3.1.1
yarl==1.9.4
zipp==3.21.0
